Abstracting access to data/configuration files
This provide an abstraction layer over the way PreludeCorrelator
access data and configuration files, depending on its installation method.

When PreludeCorrelator is installed as an EGG, data/configuration
files should be self contained within the package. We set the zip_false
flags to False so that PreludeCorrelator can expect data written to the
data files to be persistant accross run.

However, for standard prelude-correlator installation, we keep
using /etc/prelude-correlator as the configuration directory, and
/var/lib/prelude-correlator as the data directory.

This additionally introduce a default configuration file, which
installation location depend on the type of installation.