Improve WormPlugin accuracy
The plugin used to alert when seeing an alert to a given target, and
this same alert going back to the source. This can happen in a number
of case (example: Netbios alert triggered by Snort).

As of now, the plugin will wait for the events to be repeated against
at least 5 differents hosts.