Correctly set CorrelationAlert DetectTime
Modify addAlertReference() so that the reported CorrelationAlert
DetectTime match the time of the first event that was detected.

From IDMEF RFC 4765, section 4.2.2:
DetectTime

The time the event(s) leading up to the alert was detected. In the case
of more than one event, the time the first event was detected.  In some
circumstances, this may not be the same value as CreateTime.