Implement Context time merging capability
This commit introduce time merging capability within Prelude-Correlator:
as of now, correlation contexts are associated with the time of the
triggering IDMEF event.

When searching for a matching context, an existing context will only be
returned in case the specified search time match the time windows of an
existing context.

This mean that when receiving a burst of events, the time carried with
events is directly taken into account : separate instances exists for
the same Context associated with events received at a different time.

Additionally, context can be merged if more than one context time windows
overlap.