Context class improvement: overwrite, and new update method
Context initialization now take an optional 'overwrite' argument. This
argument, if set to False, mean that the Context() will be returned
un-modified if it already exist. If it doesn't, it will be created.

There is now a new update() method, which provide exactly the same
functionality as calling Context() with the 'update=True' argument.
This is useful since some plugin might want to defer an update to
another place in the code.

Remove CheckAndDecThreshold() method. This function can be replaced
by using Context.update() along with the Context.getUpdateCount()
method.

[Scan]: Update to the new Context API (removal of CheckAndDecThreshold).
Improve the plugin so that it report CorrelationAlert after the
timer expire, not after the threshold is reached.