Firewall plugin improvement
The plugin will now report CorrelationAlert for events / sets of events
that appear to have passed through a firewall known to protect the target
machine.

If no firewall ever emit block concerning a given host, then this host
is considered un-protected, and there is no point in reporting
CorrelationAlert.

The 'flush-protected-hosts' variable allow you to define how much
time a given target hosts should be considered as protected when a
firewall drop is noticed for this machine.