Ability to specify an input IDMEF criteria
There is now a [general] configuration section in prelude-correlator.conf
allowing to specify a criteria option that use an IDMEF criteria parameter
as an argument.

Using this option, it is possible to specify an input filter for events
to be processed by the correlator.

Additionally, always use an IDMEFCriteria to check whether the input
event is an alert, in place of manually retrieving alert.create_time.