Allow plugin to specify a list of conflicting plugin
Using this interface, a plugin can now provide a list of other plugin that
conflict with him, and that shouldn't be loaded.

Exemple of use:

 class MyPlugin(pluginmanager.Plugin):
    conflict = ("deprecated", ("EventScanPlugin", "EventSweepPlugin")),
               ("obsolete", ("PluginA", "PluginN"))

Additionally, provide a way for the user to force loading of a plugin
that was disabled because of a conflict (using enable = force in the
configuration section for the plugin).

Fix #229.

Change-Id: I849cec724e05ee2fbd07ce8fe5c8e85435b36ed5