Support IDMEF Alert merge
This implement support for merging different IDMEF Alert. Following
this commit, one should not see CorrelationAlert listing the same
source multiples times.

The merging is done in a smart way, for exemple two source that are
exactly the same but with different port will result in one source
with a portlist member.

Albait complex, the system should make it easy to add additional
merging features.

Change-Id: I39fe8b1b0f02d41972a55a5527c05897ed240511